fix GetNodeDouble error
Former-commit-id: 90db123e55f6eb79802d501f528591f92ebe1961
Former-commit-id: 70b33d65380733489b38979267b0bf88c653673d

+++ Frame/SDK/KernelPlugin/AFCElementModule.cpp
@@ -280,7 +280,7 @@ double AFCElementModule::GetNodeDouble(const std::string& strConfigName, const s
     AFDataNode* pNode = GetNode(strConfigName, strDataNodeName);
     if (pNode != nullptr)
     {
-        return pNode->value.GetFloat();
+        return pNode->value.GetDouble();
     }
 
     return NULL_DOUBLE;
